# Conflicts:
#	src/main/java/cn/edu/sicau/pfdistribution/service/road/KServiceImpl.java
#	src/test/java/cn/edu/sicau/pfdistribution/RoadTest.java

+++ src/main/java/cn/edu/sicau/pfdistribution/service/kspdistribution/CalculateBaseImplementation.scala
@@ -204,8 +204,7 @@ class CalculateBaseImplementation @Autowired() (val dynamicCosting:KspDynamicCos
     if(ksp == null){
       println("错误OD"+targetOd)
     }
-
-    val passengers:Int = odMap(targetOd).toInt
+    val passengers:Int = odMap(targetOd).toDouble.toInt
     val iter = ksp.iterator()
     var text:mutable.Map[Iterator[DirectedEdge], Double] = mutable.Map()
     var text1:mutable.Map[Array[DirectedEdge], Double] = mutable.Map()
